Add reservation Controller

+++ Models/Employee.cs
@@ -10,7 +10,7 @@ namespace CinemaAppBackend.Models
             Tickets = new HashSet<Ticket>();
         }
 
-        public ushort Id { get; set; }
+        public int Id { get; set; }
         public string Name { get; set; }
         public string Surname { get; set; }
         public decimal? PhoneNumber { get; set; }
