+++ Models/Seat.cs
@@ -10,10 +10,10 @@ namespace CinemaAppBackend.Models
             Tickets = new HashSet<Ticket>();
         }
 
-        public ushort Id { get; set; }
-        public byte? RoomId { get; set; }
-        public sbyte Rowing { get; set; }
-        public sbyte SeatNumber { get; set; }
+        public int Id { get; set; }
+        public int? RoomId { get; set; }
+        public int Rowing { get; set; }
+        public int SeatNumber { get; set; }
 
         public virtual Room Room { get; set; }
         public virtual ICollection<Ticket> Tickets { get; set; }
